website/requirements.txt
celery==5.0.5
dj-database-url==0.5.0
django==3.1
psycopg2-binary==2.8.6
python-decouple==3.4
redis==3.5.3
requests==2.25.1
whitenoise==5.2.0